Add "[2021-07-19] Challenge #399 [Easy] Letter value sum" solution
This solves the main task, and most of the optional ones. I can't be
bothered to solve the rest of the optional ones.

I should have solved this by mapping letter sums to collections of words
with those letter sums, but I didn't think of it until I'd already
written the solutions I've completed. Again, I have no interest in going
back and rewriting things.